+++ server/server-api/src/main/java/cn/torna/api/open/SwaggerApi.java
@@ -294,7 +294,7 @@ public class SwaggerApi {
     }
 
     private static List<DocPushItemParam> buildItems(Map.Entry<String, PathItem> entry, OpenAPI openAPI) {
-         List<DocPushItemParam> items = new ArrayList<>();
+        List<DocPushItemParam> items = new ArrayList<>();
         String path = entry.getKey();
         PathItem pathItem = entry.getValue();
         Map<PathItem.HttpMethod, Operation> operationMap = pathItem.readOperationsMap();
@@ -435,10 +435,14 @@ public class SwaggerApi {
                         if ("array".equals(type)) {
                             isResponseArray = true;
                             Schema<?> items = schema.getItems();
-                            $ref = items.get$ref();
-                            docParamPushParams = buildObjectParam($ref, openAPI, new BuildObjectParamContext());
+                            if (items != null) {
+                                $ref = items.get$ref();
+                                docParamPushParams = buildObjectParam($ref, openAPI, new BuildObjectParamContext());
+                            }
                         } else if ($ref != null && !$ref.endsWith("@")/*排除@*/) {
                             docParamPushParams = buildObjectParam($ref, openAPI, new BuildObjectParamContext());
+                        } else {
+                            docParamPushParams = buildObjectParam0(schema, openAPI, new BuildObjectParamContext());
                         }
                     }
                 }
@@ -465,6 +469,9 @@ public class SwaggerApi {
             return null;
         }
         JsonSchema jsonSchema = getJsonSchema($ref, openAPI);
+        if (jsonSchema == null) {
+            return null;
+        }
         return buildObjectParam(jsonSchema, openAPI, context);
     }
 
@@ -481,7 +488,65 @@ public class SwaggerApi {
                             .required(Booleans.toValue(jsonSchema.getRequired(name) || Objects.equals("true", String.valueOf(value.getRequired()))))
                             .description(value.getDescription())
                             .example(toString(value.getExample()))
-                            .maxLength(getMaxLength(jsonSchema.getSchema()))
+                            .maxLength(getMaxLength(jsonSchema.getSchema(), value))
+                            .build();
+                    String type = value.getType();
+                    List<DocParamPushParam> children = null;
+                    // 如果有子对象的ref
+                    if (value.get$ref() != null) {
+                        type = TYPE_OBJECT;
+                        final String child$ref = value.get$ref();
+                        children = buildObjectParam(child$ref, openAPI, context);
+                    }
+                    // 如果直接书写了子对象的内容
+                    if(value.getProperties()!=null){
+                        type = TYPE_OBJECT;
+                        children = buildObjectParam(getJsonSchema(value), openAPI, context);
+                    }
+                    // 如果是枚举字段
+                    if (value.getEnum() != null) {
+                        type = TYPE_ENUM;
+                        List<?> list = value.getEnum();
+                        setEnumDescription(list, param);
+                    }
+                    // list对象，List<XX>
+                    if (TYPE_ARRAY.equals(type) && value.getItems() != null) {
+                        Schema items = value.getItems();
+                        String itemRef = items.get$ref();
+                        if (itemRef != null) {
+                            final String child$ref = itemRef;
+                            children = buildObjectParam(child$ref, openAPI, context);
+                            type = "array[object]";
+                        }
+                        String itemType = items.getType();
+                        if (itemType != null) {
+                            type = "array[" + itemType + "]";
+                            if(TYPE_OBJECT.equals(itemType)){
+                                children = buildObjectParam(getJsonSchema(items), openAPI, context);
+                            }
+                        }
+                    }
+                    param.setChildren(children);
+                    param.setType(type);
+                    return param;
+                })
+                .collect(Collectors.toList());
+    }
+
+    private static List<DocParamPushParam> buildObjectParam0(Schema<?> schema, OpenAPI openAPI, BuildObjectParamContext context) {
+        final Map<String, Schema> properties = schema.getProperties();
+        return Optional.ofNullable(properties)
+                .orElse(Collections.emptyMap()).entrySet()
+                .stream()
+                .map(entry -> {
+                    String name = entry.getKey();
+                    Schema value = entry.getValue();
+                    DocParamPushParam param = DocParamPushParam.builder()
+                            .name(name)
+                            .required(Booleans.toValue(Objects.equals("true", String.valueOf(value.getRequired()))))
+                            .description(value.getDescription())
+                            .example(toString(value.getExample()))
+                            .maxLength(getMaxLength(schema, value))
                             .build();
                     String type = value.getType();
                     List<DocParamPushParam> children = null;
@@ -650,10 +715,17 @@ public class SwaggerApi {
     }
 
     private static String getMaxLength(Schema<?> schema) {
-        Integer maxLength = Optional.ofNullable(schema)
+        return Optional.ofNullable(schema)
                 .map(Schema::getMaxLength)
-                .orElse(null);
-        return maxLength == null ? "-" : maxLength.toString();
+                .map(String::valueOf)
+                .orElse("-");
+    }
+
+    private static String getMaxLength(Schema<?> schema, Schema<?> value) {
+        return Optional.ofNullable(schema)
+                .map(Schema::getMaxLength)
+                .map(String::valueOf)
+                .orElseGet(() -> getMaxLength(value));
     }
 
     /**
@@ -665,6 +737,9 @@ public class SwaggerApi {
     private static JsonSchema getJsonSchema(String $ref, OpenAPI openAPI) {
         String refName = getRefName($ref);
         Schema<?> schema = openAPI.getComponents().getSchemas().get(refName);
+        if (schema == null) {
+            return null;
+        }
         return getJsonSchema(schema);
     }